A block of code was used but not required
This fix removes a block of code. It was the one I had first started with to begin the process. The second block was to fix some shortcomings the first block had. I've realized now that the second block makes the first one unnecessary, so I'm removing the first one.